refactor(UserLotto): 당첨 번호 Lotto 클래스로 객체화 및 예외 처리 이동
